[BREAKING CHANGE] Require PHP 7.4 or later
In preparation for integrating wikimedia/ip-set into this library,
match its PHP requirement. We currently got away with not bumping this
because IPUtils has remained agnostic from ip-set 2.x to 4.x, but once
we merge the libraries, these need to have the same requirement.

Also bump mediawiki/mediawiki-phan-config while at it, because this
repo is currently failing Phan on local development when tested on
PHP 8.2 due to https://github.com/phan/phan/issues/4751, which was
fixed in Phan 5.4.2, and mediawiki/mediawiki-phan-config 0.13, updates
use of Phan from 5.4.1 to 5.4.2.

Change-Id: Id2ba7d59a46068a23bd30ab2717f62c8c4da0e28