Remove extra case for cidr /0 in parseCIDR
The extra case for $bits === 0 does not work since e68cf6a14e,
as $bits is still a string after is_numeric (and phan was right here).
The base_convert code works also for 0, so no need for the extra case
here. Also the /0 case seems very rare and does not need this extra.

Change-Id: Ib5545f33f27b96a854e214df3a1b4db690e5088b