Remove redundant sanitizeIPv6 call in convertIPv6ToRawHex
The internal convertIPv6ToRawHex() method's callers are now responsible
for sanitizing input. One caller (parseCIDR6) already did this, and the
other (toHex) is updated to do so. (Note that the doc-comment for
convertIPv6ToRawHex already specified this requirement.)

This makes the call within convertIPv6ToRawHex() redundant and reduces
the number of sanitize calls from four to three for an
IPUtils::isInRange() check with IPv6 inputs.

Change-Id: Ie2ab50deb62d9ed3b90f4660147605ae1923fd93